+++ src/com/yanncebron/m68kplugin/lang/psi/M68kTokenGroups.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Authors
+ * Copyright 2025 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -173,6 +173,7 @@ public final class M68kTokenGroups {
     AC68080,
     ADDWATCH,
     ALIGN,
+    ASSERT,
     AUTO,
     BLK,
     BSS,